OldBook/requirements.txt
django_phone_field==1.8.0
Django==2.2.7
channels==0.17.3
